Update [spark tests]
